+++ src/main/java/pl/pingpong/repository/PersonDataRepository.java
@@ -3,5 +3,5 @@ package pl.pingpong.repository;
 import org.springframework.data.repository.CrudRepository;
 import pl.pingpong.entity.PersonData;
 
-public interface PersonDataRepository extends CrudRepository<PersonData, Long> {
+public interface PersonDataRepository extends CrudRepository<PersonData, Integer> {
 }
